Store object observation data per-Isolate
There is a builtins object per-Context, but the observation
state should be stored per-Isolate, since multiple Contexts
may share objects and observers.

This patch implements a new strong root in Heap that simply
stores a JSObject, and exposes that object to object-observe.js
via a runtime function. object-observe.js then stores all its
state in that object.

Added a new cctest that uses multiple contexts via the API
to exercise this behavior.